Pulled rbind into a common utility file
